requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.20
databases==0.6.3
pydantic==2.2.2
python-dotenv==1.0.1
boto3==1.28.89
twilio==8.4.1
streamlit==1.30.0
requests==2.34.0
aiofiles==23.1.0
